recognition/adni-convnext-simar/requirements.txt
torch==2.4.1+cu118
torchvision==0.19.1
timm==1.0.9
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
scikit-learn==1.5.1
tqdm==4.66.5
PyYAML==6.0.2
accelerate==0.34.2
